Implemented broadcasting logic. Set up custom ws connection closing/broadcasting logic for when a ProjectUser is deleted. Updated multiple controllers to use broadcasting and additional validation. Implemented addProjectUsers controller. Fixed incorrect array item max value in addProjectUsers schema.

+++ server/src/controllers/notes.ts
@@ -17,17 +17,18 @@ export const addNote = async (ws: WebSocket, projectID: string, username: string
   if (error) throw new BadMessageError(String(error));
 
   // respond successfully with note data
-  ws.send(JSON.stringify({ action: "addNote", success: true, data: note }));
+  if (ws.readyState === WebSocket.OPEN) ws.send(JSON.stringify({ action: "addNote", success: true, data: note }));
 };
 
 // add notes (WebSocket)
+// TODO: If there are no notes for the track already, can just overwrite the old notes array with the new one
 export const addNotes = async (ws: WebSocket, projectID: string, username: string, data: any) => {
   // validate data with Joi schema
   const { error } = addNotesSchema.validate(data, { abortEarly: false });
   if (error) throw new BadMessageError(String(error));
 
   // respond successfully with data for the notes
-  ws.send(JSON.stringify({ action: "addNotes", success: true, data: notes }));
+  if (ws.readyState === WebSocket.OPEN) ws.send(JSON.stringify({ action: "addNotes", success: true, data: notes }));
 };
 
 // update note (WebSocket)
@@ -37,7 +38,7 @@ export const updateNote = async (ws: WebSocket, projectID: string, username: str
   if (error) throw new BadMessageError(String(error));
 
   // respond successfully with note data
-  ws.send(JSON.stringify({ action: "updateNote", success: true, data: note }));
+  if (ws.readyState === WebSocket.OPEN) ws.send(JSON.stringify({ action: "updateNote", success: true, data: note }));
 };
 
 // update notes (WebSocket)
@@ -47,7 +48,7 @@ export const updateNotes = async (ws: WebSocket, projectID: string, username: st
   if (error) throw new BadMessageError(String(error));
 
   // respond successfully with data for the notes
-  ws.send(JSON.stringify({ action: "updateNotes", success: true, data: notes }));
+  if (ws.readyState === WebSocket.OPEN) ws.send(JSON.stringify({ action: "updateNotes", success: true, data: notes }));
 };
 
 // delete note (WebSocket)
@@ -57,7 +58,7 @@ export const deleteNote = async (ws: WebSocket, projectID: string, username: str
   if (error) throw new BadMessageError(String(error));
 
   // respond successfully with note data
-  ws.send(JSON.stringify({ action: "deleteNote", success: true, data: note }));
+  if (ws.readyState === WebSocket.OPEN) ws.send(JSON.stringify({ action: "deleteNote", success: true, data: note }));
 };
 
 // delete notes (WebSocket)
@@ -67,7 +68,7 @@ export const deleteNotes = async (ws: WebSocket, projectID: string, username: st
   if (error) throw new BadMessageError(String(error));
 
   // respond successfully with data for the notes
-  ws.send(JSON.stringify({ action: "deleteNotes", success: true, data: notes }));
+  if (ws.readyState === WebSocket.OPEN) ws.send(JSON.stringify({ action: "deleteNotes", success: true, data: notes }));
 };
 
 // delete all notes on track (WebSocket)
@@ -77,5 +78,5 @@ export const deleteAllNotesOnTrack = async (ws: WebSocket, projectID: string, us
   if (error) throw new BadMessageError(String(error));
 
   // respond successfully with data for the notes
-  ws.send(JSON.stringify({ action: "deleteAllNotesOnTrack", success: true, data: notes }));
+  if (ws.readyState === WebSocket.OPEN) ws.send(JSON.stringify({ action: "deleteAllNotesOnTrack", success: true, data: notes }));
 };
